+++ src/core.rs
@@ -1,5 +1,6 @@
 pub mod bus;
 pub mod db;
 pub mod message;
+pub mod si;
 pub mod thread;
 pub mod value;