+++ src/core/db.rs
@@ -2,7 +2,6 @@
 
 use crate::prelude::*;
 use chrono::prelude::*;
-use rusqlite::types::FromSql;
 use rusqlite::{params, Row};
 use rusqlite::{OptionalExtension, NO_PARAMS};
 use std::path::Path;
@@ -109,17 +108,12 @@ impl Connection {
     }
 
     /// Selects the specified sensor readings within the specified period.
-    pub fn select_values<T: FromSql>(
-        &self,
-        sensor_id: &str,
-        since: &DateTime<Local>,
-    ) -> Result<Vec<(DateTime<Local>, T)>> {
+    pub fn select_readings(&self, sensor_id: &str, since: &DateTime<Local>) -> Result<Vec<Reading>> {
         self.connection()?
             // language=sql
             .prepare_cached(
                 r#"
-                -- noinspection SqlResolve @ routine/"json_extract"
-                SELECT timestamp, json_extract(value, '$.value') as value
+                SELECT timestamp, value
                 FROM readings
                 WHERE sensor_fk = ?1 AND timestamp >= ?2
                 ORDER BY timestamp
@@ -127,9 +121,7 @@ impl Connection {
             )?
             .query_map(
                 params![hash_sensor_id(sensor_id), since.timestamp_millis()],
-                |row| -> rusqlite::Result<(DateTime<Local>, T)> {
-                    Ok((Local.timestamp_millis(row.get("timestamp")?), row.get::<_, T>("value")?))
-                },
+                get_reading,
             )?
             .map(|r| r.map_err(Into::into))
             .collect()
@@ -386,8 +378,8 @@ mod tests {
             .value(Value::Counter(42))
             .timestamp(Local.timestamp_millis(1_566_424_128_000));
         message.upsert_into(&*db.connection()?)?;
-        let readings: Vec<(_, i64)> = db.select_values("test", &Local.timestamp_millis(0))?;
-        assert_eq!(readings.get(0).unwrap(), &(message.reading.timestamp, 42));
+        let readings = db.select_readings("test", &Local.timestamp_millis(0))?;
+        assert_eq!(readings.get(0).unwrap(), &message.reading);
         Ok(())
     }